@@ -1,14 +1,8 @@
 ﻿using ECO.Data;
 using Moq;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace ECO.UnitTests.Utils.Mocks
+namespace ECO.UnitTests.Utils.Mocks;
+
+public class DataTransactionMock : Mock<IDataTransaction>
 {
-    public class DataTransactionMock : Mock<IDataTransaction>
-    {
-    }
 }
