@@ -5,6 +5,6 @@ namespace ECO.Configuration;
 
 public class DataContextOptions
 {
-    public Action<IPersistenceUnitFactory, ILoggerFactory> PersistenceUnitFactoryOptions { get; set; } = (persistenceUnitFactory, loggerFactory) => { };
+    public required Action<IPersistenceUnitFactory, ILoggerFactory> PersistenceUnitFactoryOptions { get; set; } = (persistenceUnitFactory, loggerFactory) => { };
     public bool RequireTransaction { get; set; } = false;
 }
