@@ -3,14 +3,14 @@ namespace ECO;
 /// <summary>
 /// Interface that defines an Entity with an identifier of T type
 /// </summary>
-/// <typeparam name="T"></typeparam>
+/// <typeparam name="T">The type of the identifier</typeparam>
 public interface IEntity<T> : IEquatable<IEntity<T>>
 {
     #region Properties
 
     /// <summary>
     /// Identifier of the entity
-    /// </summary>
+    /// </summary>    
     T? Identity { get; }
 
     #endregion
