@@ -3,7 +3,7 @@
 /// <summary>
 /// Interface that defines a versionble Aggregate root
 /// </summary>
-/// <typeparam name="T"></typeparam>
+/// <typeparam name="T">The type of the identifier</typeparam>
 public interface IVersionableAggregateRoot<T> : IAggregateRoot<T>, IVersionableEntity<T>
 {
 
