@@ -3,7 +3,7 @@ namespace ECO;
 /// <summary>
 /// Interface that defines a versionble Entity
 /// </summary>
-/// <typeparam name="T"></typeparam>
+/// <typeparam name="T">The type of the identifier</typeparam>
 public interface IVersionableEntity<T> : IEntity<T>
 {
     #region Properties
